Added grappling hook implementation
It's kind of janky and weak. If pull force is too high it acts like a slingshot, if pull force is too weak it feels like a rope. I might turn this into two separate items, one is more of a climbing rope and  the other is more of a batman style grappling gun